Fixing the measure test to deal with the module name
Using the module name "API" allows the tests to work however, the
routes in the application do not because they expect a module called
"Api". This changes the module in the test. Eventually, we will move
these controllers out of the API module and replace the existing
controllers with them.